requirements.txt
requests==2.28.1
jupyterlab==3.4.7
PyYAML==6.0
pyspark==3.2.1
ipykernel==6.15.0
pandas==1.3.5
pymongo[srv]
python-dotenv
-e .